requirements.txt
faker==40.11.0
numpy==2.4.3
pandas==3.0.1
plotly>=5.22.0
python-dotenv==1.1.0
requests>=2.31.0
snowflake-connector-python==3.14.0
streamlit>=1.35.0
